refactoring
Former-commit-id: 395a0e16670c7bda3e2f4d00f83328ff3eee7a4d

+++ platform/src/main/java/ai/enpasos/muzero/platform/agent/rational/gumbel/SequentialHalving.java
@@ -71,7 +71,7 @@ public class SequentialHalving {
 
         int[] result = new int[phases];
         for (int p = 0; p < phases; p++) {
-            int na = (int) ((double) budget / phases /  m);
+            int na = (int) ((double) budget / phases / (double) m);
             if (p < phases - 1) {
                 result[p] = na;
                 remainingBudget -= na * m;
